refactor(viewer): make editing code more general to use with annotation
